+++ changelog.md
@@ -3,6 +3,9 @@
 Please read these notes when updating your project's `iohk-nix`
 version. There may have been changes which could break your build.
 
+## 2021-01-04
+  * Switch default nixpkgs to nixos-unstable
+
 ## 2020-11-11
    * Switch default nixpkgs to 20.09
    * `commonLib.commitIdFromGitRepo` is deprecated in favour of nixpkgs `lib.commitIdFromGitRepo`.