+++ changelog.md
@@ -5,7 +5,8 @@ version. There may have been changes which could break your build.
 
 ## 2023-04-20
   * Added `blst` library with dynamic, and static lirbaries, as well as pkg-config metadata.
-  * Added [haskell-nix](https://github.com/input-output-hk/haskell.nix) `extraPkgconfigMappings` for `blst` and `libsodium`.
+  * Renamed `blst` to `libblst`
+  * Added [haskell-nix](https://github.com/input-output-hk/haskell.nix) `extraPkgconfigMappings` for `libblst` and `libsodium`.
     This should allow us to drop the previously needed hack to map `libsodium-vrf` to `libsodium`.
 
 ## 2021-07-22